Load hoopR men's college basketball schedule

helper that loads multiple seasons from the data repo either into memory or writes it into a db using some forwarded arguments in the dots

Usage

load_mbb_schedule(
  seasons = most_recent_mbb_season(),
  ...,
  dbConnection = NULL,
  tablename = NULL
)

Arguments

seasons

A vector of 4-digit years associated with given men's college basketball seasons. (Min: 2002)

...

Additional arguments passed to an underlying function that writes the season data into a database (used by update_mbb_db()).

dbConnection

A DBIConnection object, as returned by

tablename

The name of the play by play data table within the database

Value

Returns a tibble

Examples

# \donttest{
load_mbb_schedule(2021)
#> ── MBB Schedule Information from hoopR data repository ────────── hoopR 1.8.1 ──
#> ℹ Data updated: 2022-03-27 22:08:36 UTC
#> # A tibble: 5,282 × 72
#>           id uid    date  atten…¹ timeV…² neutr…³ confe…⁴ recent start…⁵ notes…⁶
#>        <int> <chr>  <chr>   <int> <lgl>   <lgl>   <lgl>   <lgl>  <chr>   <chr>  
#>  1 401310865 s:40~… 2021…       0 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#>  2 401310867 s:40~… 2021…    8131 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#>  3 401310866 s:40~… 2021…       0 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#>  4 401310872 s:40~… 2021…    7515 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#>  5 401310871 s:40~… 2021…    6166 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#>  6 401310870 s:40~… 2021…    7519 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#>  7 401310869 s:40~… 2021…       0 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#>  8 401310875 s:40~… 2021…       0 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#>  9 401310873 s:40~… 2021…       0 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#> 10 401310874 s:40~… 2021…    3645 TRUE    TRUE    FALSE   FALSE  2021-0… event  
#> # … with 5,272 more rows, 62 more variables: notes_headline <chr>,
#> #   type.id <int>, type.abbreviation <chr>, venue.id <int>,
#> #   venue.fullName <chr>, venue.address.city <chr>, venue.address.state <chr>,
#> #   venue.capacity <int>, venue.indoor <lgl>, status.clock <dbl>,
#> #   status.displayClock <chr>, status.period <int>, status.type.id <int>,
#> #   status.type.name <chr>, status.type.state <chr>,
#> #   status.type.completed <lgl>, status.type.description <chr>, …
# }