Load hoopR men's college basketball team box scores

helper that loads multiple seasons from the data repo either into memory or writes it into a db using some forwarded arguments in the dots

Usage

load_mbb_team_box(
  seasons = most_recent_mbb_season(),
  ...,
  dbConnection = NULL,
  tablename = NULL
)

Arguments

seasons

A vector of 4-digit years associated with given men's college basketball seasons. (Min: 2003)

...

Additional arguments passed to an underlying function that writes the season data into a database (used by update_mbb_db()).

dbConnection

A DBIConnection object, as returned by

tablename

The name of the play by play data table within the database

Value

Returns a tibble

col_name	types

team_id	character

team_uid	character

team_slug	character

team_location	character

team_name	character

team_abbreviation	character

team_display_name	character

team_short_display_name	character

team_color	character

team_alternate_color	character

team_logo	character

field_goals_made_field_goals_attempted	character

field_goal_pct	character

three_point_field_goals_made_three_point_field_goals_attempted	character

three_point_field_goal_pct	character

free_throws_made_free_throws_attempted	character

free_throw_pct	character

total_rebounds	character

offensive_rebounds	character

defensive_rebounds	character

assists	character

steals	character

blocks	character

turnovers	character

team_turnovers	character

total_turnovers	character

technical_fouls	character

total_technical_fouls	character

flagrant_fouls	character

fouls	character

largest_lead	character

home_away	character

opponent_id	integer

opponent_name	character

opponent_mascot	character

opponent_abbrev	character

game_id	integer

season	integer

season_type	integer

game_date	Date

See also

Examples

# \donttest{
  load_mbb_team_box(seasons = most_recent_mbb_season())
#> ── ESPN MBB Team Boxscores from hoopR data repository ─────────── hoopR 2.0.0 ──
#> ℹ Data updated: 2023-04-02 09:22:55 UTC
#> # A tibble: 12,438 × 40
#>    team_id team_uid         team_slug  team_location team_name team_abbreviation
#>    <chr>   <chr>            <chr>      <chr>         <chr>     <chr>            
#>  1 2390    s:40~l:41~t:2390 miami-hur… Miami         Hurrican… MIA              
#>  2 41      s:40~l:41~t:41   uconn-hus… UConn         Huskies   CONN             
#>  3 2226    s:40~l:41~t:2226 florida-a… Florida Atla… Owls      FAU              
#>  4 21      s:40~l:41~t:21   san-diego… San Diego St… Aztecs    SDSU             
#>  5 5       s:40~l:41~t:5    uab-blaze… UAB           Blazers   UAB              
#>  6 249     s:40~l:41~t:249  north-tex… North Texas   Mean Gre… UNT              
#>  7 3084    s:40~l:41~t:3084 utah-vall… Utah Valley   Wolverin… UVU              
#>  8 5       s:40~l:41~t:5    uab-blaze… UAB           Blazers   UAB              
#>  9 275     s:40~l:41~t:275  wisconsin… Wisconsin     Badgers   WISC             
#> 10 249     s:40~l:41~t:249  north-tex… North Texas   Mean Gre… UNT              
#> # ℹ 12,428 more rows
#> # ℹ 34 more variables: team_display_name <chr>, team_short_display_name <chr>,
#> #   team_color <chr>, team_alternate_color <chr>, team_logo <chr>,
#> #   field_goals_made_field_goals_attempted <chr>, field_goal_pct <chr>,
#> #   three_point_field_goals_made_three_point_field_goals_attempted <chr>,
#> #   three_point_field_goal_pct <chr>,
#> #   free_throws_made_free_throws_attempted <chr>, free_throw_pct <chr>, …
# }