Men's College Basketball win probability chart data from ESPN

Men's College Basketball win probability chart data from ESPN

espn_mbb_wp(game_id)

Arguments

game_id

(Integer required): Game ID filter for querying a single game

Value

espn_mbb_wp() - A data frame with 9 variables:

game_id: character.

Referencing game ID (should be same as game_id from other functions).

play_id: character.

Referencing play ID.

period: integer.

Clock (time) left in the game.

time_left: character.

Clock (time) left in the game.

period_seconds_left: integer.

Seconds left in the period.

game_seconds_left: integer.

Seconds left in the game.

home_win_percentage: double.

The probability of the home team winning the game.

away_win_percentage: double.

The probability of the away team winning the game (calculated as 1 - home_win_percentage - tie_percentage).

tie_percentage: double.

The probability of the game ending the final period in a tie.

Examples

# \donttest{
  espn_mbb_wp(game_id = 401256760)
#>       game_id            play_id period time_left period_seconds_left
#> 1   401256760 401256760101805901      1     19:40                1180
#> 2   401256760 401256760101806301      1     19:36                1176
#> 3   401256760 401256760101806302      1     19:36                1176
#> 4   401256760 401256760101808801      1     19:11                1151
#> 5   401256760 401256760101809001      1     19:09                1149
#> 6   401256760 401256760101814501      1     18:54                1134
#> 7   401256760 401256760101814502      1     18:54                1134
#> 8   401256760 401256760101817001      1     18:29                1109
#> 9   401256760 401256760101818201      1     18:17                1097
#> 10  401256760 401256760101818202      1     18:17                1097
#> 11  401256760 401256760101818901      1     18:10                1090
#> 12  401256760 401256760101824001      1     17:59                1079
#> 13  401256760 401256760101824002      1     17:59                1079
#> 14  401256760 401256760101825001      1     17:49                1069
#> 15  401256760 401256760101825301      1     17:46                1066
#> 16  401256760 401256760101825901      1     17:40                1060
#> 17  401256760 401256760101826301      1     17:36                1056
#> 18  401256760 401256760101826302      1     17:36                1056
#> 19  401256760 401256760101828101      1     17:18                1038
#> 20  401256760 401256760101828102      1     17:18                1038
#> 21  401256760 401256760101828501      1     17:14                1034
#> 22  401256760 401256760101828701      1     17:12                1032
#> 23  401256760 401256760101829001      1     17:09                1029
#> 24  401256760 401256760101829901      1     17:00                1020
#> 25  401256760 401256760101829902      1     17:00                1020
#> 26  401256760 401256760101834301      1     16:56                1016
#> 27  401256760 401256760101834302      1     16:56                1016
#> 28  401256760 401256760101834303      1     16:56                1016
#> 29  401256760 401256760101834601      1     16:53                1013
#> 30  401256760 401256760101834602      1     16:53                1013
#> 31  401256760 401256760101834603      1     16:53                1013
#> 32  401256760 401256760101834604      1     16:53                1013
#> 33  401256760 401256760101836201      1     16:37                 997
#> 34  401256760 401256760101836501      1     16:34                 994
#> 35  401256760 401256760101837201      1     16:27                 987
#> 36  401256760 401256760101837202      1     16:27                 987
#> 37  401256760 401256760101837601      1     16:23                 983
#> 38  401256760 401256760101838101      1     16:18                 978
#> 39  401256760 401256760101838102      1     16:18                 978
#> 40  401256760 401256760101839801      1     16:01                 961
#> 41  401256760 401256760101847301      1     15:26                 926
#> 42  401256760 401256760101847302      1     15:26                 926
#> 43  401256760 401256760101854001      1     14:59                 899
#> 44  401256760 401256760101855501      1     14:44                 884
#> 45  401256760 401256760101855502      1     14:44                 884
#> 46  401256760 401256760101855901      1     14:40                 880
#> 47  401256760 401256760101856301      1     14:36                 876
#> 48  401256760 401256760101857901      1     14:20                 860
#> 49  401256760 401256760101858101      1     14:18                 858
#> 50  401256760 401256760101859501      1     14:04                 844
#> 51  401256760 401256760101859901      1     14:00                 840
#> 52  401256760 401256760101859902      1     14:00                 840
#> 53  401256760 401256760101859903      1     14:00                 840
#> 54  401256760 401256760101859904      1     14:00                 840
#> 55  401256760 401256760101864101      1     13:58                 838
#> 56  401256760 401256760101867001      1     13:29                 809
#> 57  401256760 401256760101867002      1     13:29                 809
#> 58  401256760 401256760101867003      1     13:29                 809
#> 59  401256760 401256760101867601      1     13:23                 803
#> 60  401256760 401256760101868101      1     13:18                 798
#> 61  401256760 401256760101868601      1     13:13                 793
#> 62  401256760 401256760101869001      1     13:09                 789
#> 63  401256760 401256760101874101      1     12:58                 778
#> 64  401256760 401256760101874601      1     12:53                 773
#> 65  401256760 401256760101875301      1     12:46                 766
#> 66  401256760 401256760101877201      1     12:27                 747
#> 67  401256760 401256760101877801      1     12:21                 741
#> 68  401256760 401256760101879101      1     12:08                 728
#> 69  401256760 401256760101885101      1     11:48                 708
#> 70  401256760 401256760101885401      1     11:45                 705
#> 71  401256760 401256760101885901      1     11:40                 700
#> 72  401256760 401256760101888501      1     11:14                 674
#> 73  401256760 401256760101888801      1     11:11                 671
#> 74  401256760 401256760101889102      1     11:08                 668
#> 75  401256760 401256760101889103      1     11:08                 668
#> 76  401256760 401256760101896101      1     10:38                 638
#> 77  401256760 401256760101897701      1     10:22                 622
#> 78  401256760 401256760101897702      1     10:22                 622
#> 79  401256760 401256760101897703      1     10:22                 622
#> 80  401256760 401256760101897801      1     10:21                 621
#> 81  401256760 401256760101897901      1     10:20                 620
#> 82  401256760 401256760101897902      1     10:20                 620
#> 83  401256760 401256760101897903      1     10:20                 620
#> 84  401256760 401256760101897904      1     10:20                 620
#> 85  401256760 401256760101899001      1     10:09                 609
#> 86  401256760 401256760101904501      1     09:54                 594
#> 87  401256760 401256760101904801      1     09:51                 591
#> 88  401256760 401256760101904901      1     09:50                 590
#> 89  401256760 401256760101904904      1     09:50                 590
#> 90  401256760 401256760101906801      1     09:31                 571
#> 91  401256760 401256760101907701      1     09:22                 562
#> 92  401256760 401256760101908001      1     09:19                 559
#> 93  401256760 401256760101908002      1     09:19                 559
#> 94  401256760 401256760101908003      1     09:19                 559
#> 95  401256760 401256760101908004      1     09:19                 559
#> 96  401256760 401256760101908201      1     09:17                 557
#> 97  401256760 401256760101909201      1     09:07                 547
#> 98  401256760 401256760101914701      1     08:52                 532
#> 99  401256760 401256760101915001      1     08:49                 529
#> 100 401256760 401256760101916701      1     08:32                 512
#> 101 401256760 401256760101917101      1     08:28                 508
#> 102 401256760 401256760101918201      1     08:17                 497
#> 103 401256760 401256760101918501      1     08:14                 494
#> 104 401256760 401256760101924401      1     07:55                 475
#> 105 401256760 401256760101924701      1     07:52                 472
#> 106 401256760 401256760101925202      1     07:47                 467
#> 107 401256760 401256760101925203      1     07:47                 467
#> 108 401256760 401256760101925204      1     07:47                 467
#> 109 401256760 401256760101926701      1     07:32                 452
#> 110 401256760 401256760101927001      1     07:29                 449
#> 111 401256760 401256760101929201      1     07:07                 427
#> 112 401256760 401256760101935501      1     06:44                 404
#> 113 401256760 401256760101935502      1     06:44                 404
#> 114 401256760 401256760101937601      1     06:23                 383
#> 115 401256760 401256760101938701      1     06:12                 372
#> 116 401256760 401256760101939001      1     06:09                 369
#> 117 401256760 401256760101944501      1     05:54                 354
#> 118 401256760 401256760101945601      1     05:43                 343
#> 119 401256760 401256760101945602      1     05:43                 343
#> 120 401256760 401256760101946301      1     05:36                 336
#> 121 401256760 401256760101946302      1     05:36                 336
#> 122 401256760 401256760101946303      1     05:36                 336
#> 123 401256760 401256760101948301      1     05:16                 316
#> 124 401256760 401256760101948701      1     05:12                 312
#> 125 401256760 401256760101948702      1     05:12                 312
#> 126 401256760 401256760101949101      1     05:08                 308
#> 127 401256760 401256760101954001      1     04:59                 299
#> 128 401256760 401256760101956107      1     04:38                 278
#> 129 401256760 401256760101956108      1     04:38                 278
#> 130 401256760 401256760101956109      1     04:38                 278
#> 131 401256760 401256760101956110      1     04:38                 278
#> 132 401256760 401256760101958101      1     04:18                 258
#> 133 401256760 401256760101964201      1     03:57                 237
#> 134 401256760 401256760101964701      1     03:52                 232
#> 135 401256760 401256760101966001      1     03:39                 219
#> 136 401256760 401256760101969401      1     03:05                 185
#> 137 401256760 401256760101969501      1     03:04                 184
#> 138 401256760 401256760101969502      1     03:04                 184
#> 139 401256760 401256760101969504      1     03:04                 184
#> 140 401256760 401256760101969507      1     03:04                 184
#> 141 401256760 401256760101969601      1     03:03                 183
#> 142 401256760 401256760101974401      1     02:55                 175
#> 143 401256760 401256760101974501      1     02:54                 174
#> 144 401256760 401256760101974504      1     02:54                 174
#> 145 401256760 401256760101974506      1     02:54                 174
#> 146 401256760 401256760101974507      1     02:54                 174
#> 147 401256760 401256760101974701      1     02:52                 172
#> 148 401256760 401256760101975801      1     02:41                 161
#> 149 401256760 401256760101975802      1     02:41                 161
#> 150 401256760 401256760101977201      1     02:27                 147
#> 151 401256760 401256760101977501      1     02:24                 144
#> 152 401256760 401256760101977502      1     02:24                 144
#> 153 401256760 401256760101977503      1     02:24                 144
#> 154 401256760 401256760101977504      1     02:24                 144
#> 155 401256760 401256760101977505      1     02:24                 144
#> 156 401256760 401256760101979501      1     02:04                 124
#> 157 401256760 401256760101979601      1     02:03                 123
#> 158 401256760 401256760101984301      1     01:56                 116
#> 159 401256760 401256760101986001      1     01:39                  99
#> 160 401256760 401256760101988301      1     01:16                  76
#> 161 401256760 401256760101989401      1     01:05                  65
#> 162 401256760 401256760101995001      1     00:49                  49
#> 163 401256760 401256760101995901      1     00:40                  40
#> 164 401256760 401256760101995902      1     00:40                  40
#> 165 401256760 401256760101995903      1     00:40                  40
#> 166 401256760 401256760101998701      1     00:12                  12
#> 167 401256760 401256760101998801      1     00:11                  11
#> 168 401256760 401256760101998802      1     00:11                  11
#> 169 401256760 401256760101998803      1     00:11                  11
#> 170 401256760 401256760101998804      1     00:11                  11
#> 171 401256760 401256760101999901      1     00:00                   0
#> 172 401256760 401256760101999902      1     00:00                   0
#> 173 401256760 401256760101999903      1     00:00                   0
#> 174 401256760 401256760102806201      2     19:37                1177
#> 175 401256760 401256760102806202      2     19:37                1177
#> 176 401256760 401256760102806701      2     19:32                1172
#> 177 401256760 401256760102806801      2     19:31                1171
#> 178 401256760 401256760102806802      2     19:31                1171
#> 179 401256760 401256760102806901      2     19:30                1170
#> 180 401256760 401256760102807101      2     19:28                1168
#> 181 401256760 401256760102808201      2     19:17                1157
#> 182 401256760 401256760102808601      2     19:13                1153
#> 183 401256760 401256760102809301      2     19:06                1146
#> 184 401256760 401256760102814001      2     18:59                1139
#> 185 401256760 401256760102816101      2     18:38                1118
#> 186 401256760 401256760102817101      2     18:28                1108
#> 187 401256760 401256760102818201      2     18:17                1097
#> 188 401256760 401256760102819601      2     18:03                1083
#> 189 401256760 401256760102824601      2     17:53                1073
#> 190 401256760 401256760102824602      2     17:53                1073
#> 191 401256760 401256760102825001      2     17:49                1069
#> 192 401256760 401256760102825002      2     17:49                1069
#> 193 401256760 401256760102826201      2     17:37                1057
#> 194 401256760 401256760102826501      2     17:34                1054
#> 195 401256760 401256760102828001      2     17:19                1039
#> 196 401256760 401256760102829801      2     17:01                1021
#> 197 401256760 401256760102829802      2     17:01                1021
#> 198 401256760 401256760102835601      2     16:43                1003
#> 199 401256760 401256760102837801      2     16:21                 981
#> 200 401256760 401256760102837802      2     16:21                 981
#> 201 401256760 401256760102839201      2     16:07                 967
#> 202 401256760 401256760102839401      2     16:05                 965
#> 203 401256760 401256760102844401      2     15:55                 955
#> 204 401256760 401256760102845101      2     15:48                 948
#> 205 401256760 401256760102845102      2     15:48                 948
#> 206 401256760 401256760102845801      2     15:41                 941
#> 207 401256760 401256760102846101      2     15:38                 938
#> 208 401256760 401256760102847001      2     15:29                 929
#> 209 401256760 401256760102847301      2     15:26                 926
#> 210 401256760 401256760102848401      2     15:15                 915
#> 211 401256760 401256760102848402      2     15:15                 915
#> 212 401256760 401256760102848403      2     15:15                 915
#> 213 401256760 401256760102849901      2     15:00                 900
#> 214 401256760 401256760102855001      2     14:49                 889
#> 215 401256760 401256760102855002      2     14:49                 889
#> 216 401256760 401256760102855501      2     14:44                 884
#> 217 401256760 401256760102858701      2     14:12                 852
#> 218 401256760 401256760102859001      2     14:09                 849
#> 219 401256760 401256760102859301      2     14:06                 846
#> 220 401256760 401256760102859302      2     14:06                 846
#> 221 401256760 401256760102859701      2     14:02                 842
#> 222 401256760 401256760102864301      2     13:56                 836
#> 223 401256760 401256760102866901      2     13:30                 810
#> 224 401256760 401256760102866902      2     13:30                 810
#> 225 401256760 401256760102867401      2     13:25                 805
#> 226 401256760 401256760102867402      2     13:25                 805
#> 227 401256760 401256760102867403      2     13:25                 805
#> 228 401256760 401256760102868601      2     13:13                 793
#> 229 401256760 401256760102874501      2     12:54                 774
#> 230 401256760 401256760102874502      2     12:54                 774
#> 231 401256760 401256760102875601      2     12:43                 763
#> 232 401256760 401256760102876101      2     12:38                 758
#> 233 401256760 401256760102876501      2     12:34                 754
#> 234 401256760 401256760102878201      2     12:17                 737
#> 235 401256760 401256760102878301      2     12:16                 736
#> 236 401256760 401256760102879201      2     12:07                 727
#> 237 401256760 401256760102879901      2     12:00                 720
#> 238 401256760 401256760102879902      2     12:00                 720
#> 239 401256760 401256760102885401      2     11:45                 705
#> 240 401256760 401256760102887201      2     11:27                 687
#> 241 401256760 401256760102889701      2     11:02                 662
#> 242 401256760 401256760102889702      2     11:02                 662
#> 243 401256760 401256760102889901      2     11:00                 660
#> 244 401256760 401256760102889902      2     11:00                 660
#> 245 401256760 401256760102897101      2     10:28                 628
#> 246 401256760 401256760102897102      2     10:28                 628
#> 247 401256760 401256760102897104      2     10:28                 628
#> 248 401256760 401256760102897401      2     10:25                 625
#> 249 401256760 401256760102897402      2     10:25                 625
#> 250 401256760 401256760102897501      2     10:24                 624
#> 251 401256760 401256760102898701      2     10:12                 612
#> 252 401256760 401256760102899901      2     10:00                 600
#> 253 401256760 401256760102905401      2     09:45                 585
#> 254 401256760 401256760102905801      2     09:41                 581
#> 255 401256760 401256760102906201      2     09:37                 577
#> 256 401256760 401256760102906601      2     09:33                 573
#> 257 401256760 401256760102908901      2     09:10                 550
#> 258 401256760 401256760102908902      2     09:10                 550
#> 259 401256760 401256760102908903      2     09:10                 550
#> 260 401256760 401256760102909801      2     09:01                 541
#> 261 401256760 401256760102914101      2     08:58                 538
#> 262 401256760 401256760102915101      2     08:48                 528
#> 263 401256760 401256760102916101      2     08:38                 518
#> 264 401256760 401256760102916102      2     08:38                 518
#> 265 401256760 401256760102916103      2     08:38                 518
#> 266 401256760 401256760102917401      2     08:25                 505
#> 267 401256760 401256760102917901      2     08:20                 500
#> 268 401256760 401256760102919601      2     08:03                 483
#> 269 401256760 401256760102919602      2     08:03                 483
#> 270 401256760 401256760102924701      2     07:52                 472
#> 271 401256760 401256760102926501      2     07:34                 454
#> 272 401256760 401256760102926502      2     07:34                 454
#> 273 401256760 401256760102928501      2     07:14                 434
#> 274 401256760 401256760102929901      2     07:00                 420
#> 275 401256760 401256760102929902      2     07:00                 420
#> 276 401256760 401256760102929903      2     07:00                 420
#> 277 401256760 401256760102935301      2     06:46                 406
#> 278 401256760 401256760102935302      2     06:46                 406
#> 279 401256760 401256760102936101      2     06:38                 398
#> 280 401256760 401256760102936301      2     06:36                 396
#> 281 401256760 401256760102936302      2     06:36                 396
#> 282 401256760 401256760102936303      2     06:36                 396
#> 283 401256760 401256760102936401      2     06:35                 395
#> 284 401256760 401256760102938201      2     06:17                 377
#> 285 401256760 401256760102938901      2     06:10                 370
#> 286 401256760 401256760102938902      2     06:10                 370
#> 287 401256760 401256760102939301      2     06:06                 366
#> 288 401256760 401256760102939501      2     06:04                 364
#> 289 401256760 401256760102945701      2     05:42                 342
#> 290 401256760 401256760102946101      2     05:38                 338
#> 291 401256760 401256760102947101      2     05:28                 328
#> 292 401256760 401256760102948601      2     05:13                 313
#> 293 401256760 401256760102948602      2     05:13                 313
#> 294 401256760 401256760102948901      2     05:10                 310
#> 295 401256760 401256760102949301      2     05:06                 306
#> 296 401256760 401256760102949302      2     05:06                 306
#> 297 401256760 401256760102949303      2     05:06                 306
#> 298 401256760 401256760102954701      2     04:52                 292
#> 299 401256760 401256760102955101      2     04:48                 288
#> 300 401256760 401256760102956301      2     04:36                 276
#> 301 401256760 401256760102957004      2     04:29                 269
#> 302 401256760 401256760102957201      2     04:27                 267
#> 303 401256760 401256760102958601      2     04:13                 253
#> 304 401256760 401256760102964401      2     03:55                 235
#> 305 401256760 401256760102964402      2     03:55                 235
#> 306 401256760 401256760102964403      2     03:55                 235
#> 307 401256760 401256760102964404      2     03:55                 235
#> 308 401256760 401256760102964405      2     03:55                 235
#> 309 401256760 401256760102964406      2     03:55                 235
#> 310 401256760 401256760102965201      2     03:47                 227
#> 311 401256760 401256760102965601      2     03:43                 223
#> 312 401256760 401256760102965701      2     03:42                 222
#> 313 401256760 401256760102968401      2     03:15                 195
#> 314 401256760 401256760102968701      2     03:12                 192
#> 315 401256760 401256760102969201      2     03:07                 187
#> 316 401256760 401256760102976401      2     02:35                 155
#> 317 401256760 401256760102976801      2     02:31                 151
#> 318 401256760 401256760102976802      2     02:31                 151
#> 319 401256760 401256760102976803      2     02:31                 151
#> 320 401256760 401256760102977301      2     02:26                 146
#> 321 401256760 401256760102977302      2     02:26                 146
#> 322 401256760 401256760102977303      2     02:26                 146
#> 323 401256760 401256760102978801      2     02:11                 131
#> 324 401256760 401256760102978802      2     02:11                 131
#> 325 401256760 401256760102978803      2     02:11                 131
#> 326 401256760 401256760102978804      2     02:11                 131
#> 327 401256760 401256760102979901      2     02:00                 120
#> 328 401256760 401256760102979902      2     02:00                 120
#> 329 401256760 401256760102984101      2     01:58                 118
#> 330 401256760 401256760102985001      2     01:49                 109
#> 331 401256760 401256760102985002      2     01:49                 109
#> 332 401256760 401256760102986701      2     01:32                  92
#> 333 401256760 401256760102986702      2     01:32                  92
#> 334 401256760 401256760102987101      2     01:28                  88
#> 335 401256760 401256760102988401      2     01:15                  75
#> 336 401256760 401256760102988402      2     01:15                  75
#> 337 401256760 401256760102988403      2     01:15                  75
#> 338 401256760 401256760102994701      2     00:52                  52
#> 339 401256760 401256760102994901      2     00:50                  50
#> 340 401256760 401256760102995301      2     00:46                  46
#> 341 401256760 401256760102995302      2     00:46                  46
#> 342 401256760 401256760102995303      2     00:46                  46
#> 343 401256760 401256760102996002      2     00:39                  39
#> 344 401256760 401256760102996701      2     00:32                  32
#> 345 401256760 401256760102996702      2     00:32                  32
#> 346 401256760 401256760102997001      2     00:29                  29
#> 347 401256760 401256760102997002      2     00:29                  29
#> 348 401256760 401256760102997003      2     00:29                  29
#> 349 401256760 401256760102997801      2     00:21                  21
#> 350 401256760 401256760102998101      2     00:18                  18
#> 351 401256760 401256760102999901      2     00:00                   0
#> 352 401256760 401256760102999902      2     00:00                   0
#>     game_seconds_left home_win_percentage away_win_percentage tie_percentage
#> 1                2380               0.413               0.587              0
#> 2                2376               0.431               0.569              0
#> 3                2376               0.458               0.542              0
#> 4                2351               0.482               0.518              0
#> 5                2349               0.489               0.511              0
#> 6                2334               0.467               0.533              0
#> 7                2334               0.459               0.541              0
#> 8                2309               0.409               0.591              0
#> 9                2297               0.387               0.613              0
#> 10               2297               0.387               0.613              0
#> 11               2290               0.320               0.680              0
#> 12               2279               0.304               0.696              0
#> 13               2279               0.304               0.696              0
#> 14               2269               0.316               0.684              0
#> 15               2266               0.321               0.679              0
#> 16               2260               0.308               0.692              0
#> 17               2256               0.304               0.696              0
#> 18               2256               0.321               0.679              0
#> 19               2238               0.308               0.692              0
#> 20               2238               0.308               0.692              0
#> 21               2234               0.304               0.696              0
#> 22               2232               0.317               0.683              0
#> 23               2229               0.322               0.678              0
#> 24               2220               0.304               0.696              0
#> 25               2220               0.304               0.696              0
#> 26               2216               0.277               0.723              0
#> 27               2216               0.257               0.743              0
#> 28               2216               0.251               0.749              0
#> 29               2213               0.256               0.744              0
#> 30               2213               0.254               0.746              0
#> 31               2213               0.249               0.751              0
#> 32               2213               0.255               0.745              0
#> 33               2197               0.271               0.729              0
#> 34               2194               0.277               0.723              0
#> 35               2187               0.255               0.745              0
#> 36               2187               0.255               0.745              0
#> 37               2183               0.230               0.770              0
#> 38               2178               0.230               0.770              0
#> 39               2178               0.230               0.770              0
#> 40               2161               0.277               0.723              0
#> 41               2126               0.296               0.704              0
#> 42               2126               0.297               0.703              0
#> 43               2099               0.358               0.642              0
#> 44               2084               0.387               0.613              0
#> 45               2084               0.387               0.613              0
#> 46               2080               0.365               0.635              0
#> 47               2076               0.387               0.613              0
#> 48               2060               0.365               0.635              0
#> 49               2058               0.358               0.642              0
#> 50               2044               0.387               0.613              0
#> 51               2040               0.390               0.610              0
#> 52               2040               0.389               0.611              0
#> 53               2040               0.383               0.617              0
#> 54               2040               0.391               0.609              0
#> 55               2038               0.358               0.642              0
#> 56               2009               0.340               0.660              0
#> 57               2009               0.334               0.666              0
#> 58               2009               0.340               0.660              0
#> 59               2003               0.352               0.648              0
#> 60               1998               0.359               0.641              0
#> 61               1993               0.340               0.660              0
#> 62               1989               0.334               0.666              0
#> 63               1978               0.353               0.647              0
#> 64               1973               0.334               0.666              0
#> 65               1966               0.334               0.666              0
#> 66               1947               0.331               0.669              0
#> 67               1941               0.357               0.643              0
#> 68               1928               0.363               0.637              0
#> 69               1908               0.340               0.660              0
#> 70               1905               0.334               0.666              0
#> 71               1900               0.278               0.722              0
#> 72               1874               0.260               0.740              0
#> 73               1871               0.254               0.746              0
#> 74               1868               0.278               0.722              0
#> 75               1868               0.278               0.722              0
#> 76               1838               0.254               0.746              0
#> 77               1822               0.251               0.749              0
#> 78               1822               0.230               0.770              0
#> 79               1822               0.250               0.750              0
#> 80               1821               0.226               0.774              0
#> 81               1820               0.223               0.777              0
#> 82               1820               0.224               0.776              0
#> 83               1820               0.228               0.772              0
#> 84               1820               0.221               0.779              0
#> 85               1809               0.201               0.799              0
#> 86               1794               0.216               0.784              0
#> 87               1791               0.216               0.784              0
#> 88               1790               0.201               0.799              0
#> 89               1790               0.221               0.779              0
#> 90               1771               0.247               0.753              0
#> 91               1762               0.272               0.728              0
#> 92               1759               0.274               0.726              0
#> 93               1759               0.274               0.726              0
#> 94               1759               0.268               0.732              0
#> 95               1759               0.249               0.751              0
#> 96               1757               0.249               0.751              0
#> 97               1747               0.199               0.801              0
#> 98               1732               0.201               0.799              0
#> 99               1729               0.250               0.750              0
#> 100              1712               0.268               0.732              0
#> 101              1708               0.275               0.725              0
#> 102              1697               0.276               0.724              0
#> 103              1694               0.335               0.665              0
#> 104              1675               0.335               0.665              0
#> 105              1672               0.364               0.636              0
#> 106              1667               0.368               0.632              0
#> 107              1667               0.377               0.623              0
#> 108              1667               0.424               0.576              0
#> 109              1652               0.422               0.578              0
#> 110              1649               0.429               0.571              0
#> 111              1627               0.467               0.533              0
#> 112              1604               0.515               0.485              0
#> 113              1604               0.499               0.501              0
#> 114              1583               0.470               0.530              0
#> 115              1572               0.493               0.507              0
#> 116              1569               0.501               0.499              0
#> 117              1554               0.544               0.456              0
#> 118              1543               0.572               0.428              0
#> 119              1543               0.572               0.428              0
#> 120              1536               0.597               0.403              0
#> 121              1536               0.606               0.394              0
#> 122              1536               0.614               0.386              0
#> 123              1516               0.613               0.387              0
#> 124              1512               0.636               0.364              0
#> 125              1512               0.636               0.364              0
#> 126              1508               0.643               0.357              0
#> 127              1499               0.712               0.288              0
#> 128              1478               0.676               0.324              0
#> 129              1478               0.650               0.350              0
#> 130              1478               0.642               0.358              0
#> 131              1478               0.642               0.358              0
#> 132              1458               0.713               0.287              0
#> 133              1437               0.734               0.266              0
#> 134              1432               0.716               0.284              0
#> 135              1419               0.681               0.319              0
#> 136              1385               0.665               0.335              0
#> 137              1384               0.706               0.294              0
#> 138              1384               0.706               0.294              0
#> 139              1384               0.746               0.254              0
#> 140              1384               0.688               0.312              0
#> 141              1383               0.747               0.253              0
#> 142              1375               0.741               0.259              0
#> 143              1374               0.723               0.277              0
#> 144              1374               0.689               0.311              0
#> 145              1374               0.723               0.277              0
#> 146              1374               0.692               0.308              0
#> 147              1372               0.714               0.286              0
#> 148              1361               0.668               0.332              0
#> 149              1361               0.691               0.309              0
#> 150              1347               0.711               0.289              0
#> 151              1344               0.693               0.307              0
#> 152              1344               0.693               0.307              0
#> 153              1344               0.695               0.305              0
#> 154              1344               0.695               0.305              0
#> 155              1344               0.688               0.312              0
#> 156              1324               0.671               0.329              0
#> 157              1323               0.665               0.335              0
#> 158              1316               0.582               0.418              0
#> 159              1299               0.629               0.371              0
#> 160              1276               0.657               0.343              0
#> 161              1265               0.702               0.298              0
#> 162              1249               0.659               0.341              0
#> 163              1240               0.705               0.295              0
#> 164              1240               0.731               0.269              0
#> 165              1240               0.740               0.260              0
#> 166              1212               0.754               0.246              0
#> 167              1211               0.754               0.246              0
#> 168              1211               0.797               0.203              0
#> 169              1211               0.804               0.196              0
#> 170              1211               0.811               0.189              0
#> 171              1200               0.805               0.195              0
#> 172              1200               0.811               0.189              0
#> 173              1200               0.816               0.184              0
#> 174              1177               0.811               0.189              0
#> 175              1177               0.811               0.189              0
#> 176              1172               0.832               0.168              0
#> 177              1171               0.836               0.164              0
#> 178              1171               0.857               0.143              0
#> 179              1170               0.837               0.163              0
#> 180              1168               0.837               0.163              0
#> 181              1157               0.854               0.146              0
#> 182              1153               0.860               0.140              0
#> 183              1146               0.866               0.134              0
#> 184              1139               0.894               0.106              0
#> 185              1118               0.849               0.151              0
#> 186              1108               0.877               0.123              0
#> 187              1097               0.878               0.122              0
#> 188              1083               0.829               0.171              0
#> 189              1073               0.810               0.190              0
#> 190              1073               0.810               0.190              0
#> 191              1069               0.803               0.197              0
#> 192              1069               0.788               0.212              0
#> 193              1057               0.818               0.182              0
#> 194              1054               0.826               0.174              0
#> 195              1039               0.856               0.144              0
#> 196              1021               0.879               0.121              0
#> 197              1021               0.879               0.121              0
#> 198              1003               0.905               0.095              0
#> 199               981               0.922               0.078              0
#> 200               981               0.922               0.078              0
#> 201               967               0.914               0.086              0
#> 202               965               0.910               0.090              0
#> 203               955               0.910               0.090              0
#> 204               948               0.926               0.074              0
#> 205               948               0.926               0.074              0
#> 206               941               0.917               0.083              0
#> 207               938               0.914               0.086              0
#> 208               929               0.925               0.075              0
#> 209               926               0.929               0.071              0
#> 210               915               0.920               0.080              0
#> 211               915               0.920               0.080              0
#> 212               915               0.930               0.070              0
#> 213               900               0.959               0.041              0
#> 214               889               0.968               0.032              0
#> 215               889               0.968               0.032              0
#> 216               884               0.976               0.024              0
#> 217               852               0.982               0.018              0
#> 218               849               0.983               0.017              0
#> 219               846               0.979               0.021              0
#> 220               846               0.979               0.021              0
#> 221               842               0.978               0.022              0
#> 222               836               0.972               0.028              0
#> 223               810               0.967               0.033              0
#> 224               810               0.967               0.033              0
#> 225               805               0.961               0.039              0
#> 226               805               0.956               0.044              0
#> 227               805               0.953               0.047              0
#> 228               793               0.941               0.059              0
#> 229               774               0.953               0.047              0
#> 230               774               0.953               0.047              0
#> 231               763               0.958               0.042              0
#> 232               758               0.968               0.032              0
#> 233               754               0.963               0.037              0
#> 234               737               0.971               0.029              0
#> 235               736               0.964               0.036              0
#> 236               727               0.979               0.021              0
#> 237               720               0.973               0.027              0
#> 238               720               0.973               0.027              0
#> 239               705               0.965               0.035              0
#> 240               687               0.969               0.031              0
#> 241               662               0.962               0.038              0
#> 242               662               0.969               0.031              0
#> 243               660               0.963               0.037              0
#> 244               660               0.960               0.040              0
#> 245               628               0.955               0.045              0
#> 246               628               0.962               0.038              0
#> 247               628               0.964               0.036              0
#> 248               625               0.971               0.029              0
#> 249               625               0.971               0.029              0
#> 250               624               0.965               0.035              0
#> 251               612               0.942               0.058              0
#> 252               600               0.927               0.073              0
#> 253               585               0.941               0.059              0
#> 254               581               0.929               0.071              0
#> 255               577               0.926               0.074              0
#> 256               573               0.943               0.057              0
#> 257               550               0.948               0.052              0
#> 258               550               0.958               0.042              0
#> 259               550               0.961               0.039              0
#> 260               541               0.969               0.031              0
#> 261               538               0.971               0.029              0
#> 262               528               0.980               0.020              0
#> 263               518               0.973               0.027              0
#> 264               518               0.963               0.037              0
#> 265               518               0.961               0.039              0
#> 266               505               0.964               0.036              0
#> 267               500               0.954               0.046              0
#> 268               483               0.963               0.037              0
#> 269               483               0.966               0.034              0
#> 270               472               0.984               0.016              0
#> 271               454               0.985               0.015              0
#> 272               454               0.984               0.016              0
#> 273               434               0.990               0.010              0
#> 274               420               0.988               0.012              0
#> 275               420               0.988               0.012              0
#> 276               420               0.990               0.010              0
#> 277               406               0.988               0.012              0
#> 278               406               0.988               0.012              0
#> 279               398               0.991               0.009              0
#> 280               396               0.992               0.008              0
#> 281               396               0.992               0.008              0
#> 282               396               0.992               0.008              0
#> 283               395               0.989               0.011              0
#> 284               377               0.977               0.023              0
#> 285               370               0.964               0.036              0
#> 286               370               0.969               0.031              0
#> 287               366               0.976               0.024              0
#> 288               364               0.978               0.022              0
#> 289               342               0.973               0.027              0
#> 290               338               0.962               0.038              0
#> 291               328               0.974               0.026              0
#> 292               313               0.982               0.018              0
#> 293               313               0.982               0.018              0
#> 294               310               0.984               0.016              0
#> 295               306               0.985               0.015              0
#> 296               306               0.988               0.012              0
#> 297               306               0.990               0.010              0
#> 298               292               0.993               0.007              0
#> 299               288               0.994               0.006              0
#> 300               276               0.993               0.007              0
#> 301               269               0.992               0.008              0
#> 302               267               0.988               0.012              0
#> 303               253               0.994               0.006              0
#> 304               235               0.997               0.003              0
#> 305               235               0.997               0.003              0
#> 306               235               0.997               0.003              0
#> 307               235               0.997               0.003              0
#> 308               235               0.998               0.002              0
#> 309               235               0.999               0.001              0
#> 310               227               0.999               0.001              0
#> 311               223               0.999               0.001              0
#> 312               222               0.997               0.003              0
#> 313               195               0.997               0.003              0
#> 314               192               0.996               0.004              0
#> 315               187               0.989               0.011              0
#> 316               155               0.989               0.011              0
#> 317               151               0.993               0.007              0
#> 318               151               0.990               0.010              0
#> 319               151               0.989               0.011              0
#> 320               146               0.985               0.015              0
#> 321               146               0.982               0.018              0
#> 322               146               0.980               0.020              0
#> 323               131               0.991               0.009              0
#> 324               131               0.994               0.006              0
#> 325               131               0.991               0.009              0
#> 326               131               0.991               0.009              0
#> 327               120               0.997               0.003              0
#> 328               120               0.996               0.004              0
#> 329               118               0.993               0.007              0
#> 330               109               0.997               0.003              0
#> 331               109               0.997               0.003              0
#> 332                92               0.996               0.004              0
#> 333                92               0.996               0.004              0
#> 334                88               0.996               0.004              0
#> 335                75               0.995               0.005              0
#> 336                75               0.997               0.003              0
#> 337                75               0.999               0.001              0
#> 338                52               0.999               0.001              0
#> 339                50               0.999               0.001              0
#> 340                46               0.997               0.003              0
#> 341                46               0.997               0.003              0
#> 342                46               0.997               0.003              0
#> 343                39               0.995               0.005              0
#> 344                32               0.999               0.001              0
#> 345                32               0.999               0.001              0
#> 346                29               0.999               0.001              0
#> 347                29               0.999               0.001              0
#> 348                29               0.999               0.001              0
#> 349                21               0.999               0.001              0
#> 350                18               0.999               0.001              0
#> 351                 0               1.000               0.000              0
#> 352                 0               1.000               0.000              0
# }