Get KenPom's conference-wide stats

Get KenPom's conference-wide stats

Usage

kp_conf(year, conf)

Arguments

year

Year (YYYY)

conf

Used to limit to players in a specific conference. Allowed values are: 'A10', 'ACC', 'AE', 'AMER',
'ASUN', 'B10', 'B12', 'BE', 'BSKY', 'BSTH', 'BW', 'CAA', 'CUSA', 'HORZ', 'IND', IVY',
'MAAC', 'MAC', 'MEAC', 'MVC', 'MWC', 'NEC', 'OVC', 'P12', 'PAT', 'SB', 'SC', 'SEC', 'SLND',
'SUM', 'SWAC', 'WAC', 'WCC'.
If you try to use a conference that doesn't exist for a given season, like 'IND' and '2018',
you'll get an empty table, as kenpom.com doesn't serve 404 pages for invalid table queries like that.
No filter applied by default.

Value

A list of 7 dataframes accessible via [[1]],[[2]],...,,[[7]]

First data frame, accessible via [[1]]

A data frame with 15 columns:

Team

character.

Overall

character.

Conf

character.

AdjEM

double.

AdjEM.Rk

double.

AdjO

double.

AdjO.Rk

double.

AdjD

double.

AdjD.Rk

double.

AdjT

double.

AdjT.Rk

double.

ConfSOS

double.

ConfSOS.Rk

double.

NextGame

character.

Year

character.

Second data frame, accessible via [[2]]

A data frame with 20 columns:

Team

character.

OE

double.

OE.Rk

double.

eFG.Pct

double.

eFG.Pct.Rk

double.

TO.Pct

double.

TO.Pct.Rk

double.

OR.Pct

double.

OR.Pct.Rk

double.

FTR

double.

FTR.Rk

double.

FG2.Pct

double.

FG2.Pct.Rk

double.

FG3.Pct

double.

FG3.Pct.Rk

double.

FT.Pct

double.

FT.Pct.Rk

double.

Tempo

double.

Tempo.Rk

double.

Year

character.

Third data frame, accessible via [[3]]

A data frame with 20 columns:

Team

character.

DE

double.

DE.Rk

double.

eFG.Pct

double.

eFG.Pct.Rk

double.

TO.Pct

double.

TO.Pct.Rk

double.

OR.Pct

double.

OR.Pct.Rk

double.

FTR

double.

FTR.Rk

double.

FG2.Pct

double.

FG2.Pct.Rk

double.

FG3.Pct

double.

FG3.Pct.Rk

double.

Blk.Pct

double.

Blk.Pct.Rk

double.

Stl.Pct

double.

Stl.Pct.Rk

double.

Year

character.

Fourth data frame, accessible via [[4]]

A data frame with 3 columns:

Rk

integer.

Player

character.

Year

character.

Fifth data frame, accessible via [[5]]

A data frame with 4 columns:

Stat

character.

Value

double.

Rk

double.

Year

character.

Sixth data frame, accessible via [[6]]

A data frame with 5 columns:

Stat

character.

Count

character.

Value

double.

Rk

double.

Year

character.

Seventh data frame, accessible via [[7]]

A data frame 4 columns:

Rk

double.

Conference

character.

Rating

double.

Year

character.

Examples

# \donttest{
    try(kp_conf(year = 2020, conf = 'ACC'))
#> 2022-12-13 03:07:47: Invalid arguments or no conference data for 2020 ACC available!
#> Error in kp_conf(year = 2020, conf = "ACC") : object 'kenpom' not found
# }