Get KenPom's historical conference ratings

Get KenPom's historical conference ratings

Usage

kp_confhistory(conf)

Arguments

conf

Used to limit to players in a specific conference.

Allowed values are: 'A10', 'ACC', 'AE', 'AMER', 'ASUN', 'B10', 'B12', 'BE', 'BSKY', 'BSTH', 'BW', 'CAA', 'CUSA', 'HORZ', 'IND', IVY', 'MAAC', 'MAC', 'MEAC', 'MVC', 'MWC', 'NEC', 'OVC', 'P12', 'PAT', 'SB', 'SC', 'SEC', 'SLND', 'SUM', 'SWAC', 'WAC', 'WCC'.

If you try to use a conference that doesn't exist for a given season, like 'IND' and '2018', you'll get an empty table, as kenpom.com doesn't serve 404 pages for invalid table queries like that. No filter applied by default.

Value

A data frame with 23 columns:

Year- integer

Rank- character

Tempo- double

Efficiency- double

eFG.Pct- double

TO.Pct- double

OR.Pct- double

FTR- double

FG2.Pct- double

FG3.Pct- double

FT.Pct- double

FG3A.Pct- double

A.Pct- double

Blk.Pct- double

Stl.Pct- double

HomeRecord- character

Bids- character

S16- character

F4- character

CH- character

RegSeasonChamp- character

TourneyChamp- character

BestTeam- character

Examples

# \donttest{
 try(kp_confhistory(conf = 'ACC'))
#>    year                                                   rank tempo efficiency
#> 1  2022                                                      5  66.2      105.9
#> 2  2021                                                      5  67.2      104.1
#> 3  2020                                                      4  68.0      100.9
#> 4  2019                                                      3  67.3      102.2
#> 5  2018                                                      2  67.2      105.0
#> 6  2017                                                      2  68.0      107.9
#> 7  2016                                                      2  66.4      106.0
#> 8  2015                                                      3  63.3      104.2
#> 9    NA       Dropped: Maryland (B10) Added: Louisville (Amer)    NA         NA
#> 10 2014                                                      3  61.8      105.2
#> 11   NA Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)    NA         NA
#> 12 2013                                                      6  65.5      101.6
#> 13 2012                                                      5  65.1      100.2
#> 14 2011                                                      3  67.2      101.7
#> 15 2010                                                      2  67.8      100.5
#> 16 2009                                                      5  69.8      104.2
#> 17 2008                                                      3  71.0      104.8
#> 18 2007                                                      1  68.1      105.8
#> 19 2006                                                      3  68.6      104.7
#> 20   NA                             Added: Boston College (BE)    NA         NA
#> 21 2005                                                      1  70.6      104.8
#> 22   NA               Added: Miami FL (BE), Virginia Tech (BE)    NA         NA
#> 23 2004                                                      1  70.1      103.0
#> 24 2003                                                      3  69.6      102.5
#> 25 2002                                                      3  74.2      106.3
#> 26 2001                                                      1  73.7      103.7
#> 27 2000                                                      3  70.0      102.2
#> 28 1999                                                      2  69.8      101.7
#> 29 1998                                                      1  68.7      102.5
#> 30 1997                                                      1  65.1      100.2
#>    e_fg_pct to_pct or_pct ft_rate fg_2_pct fg_3_pct ft_pct fg_3a_pct a_pct
#> 1      51.4   16.7   27.7    27.0     50.6     35.1   73.8      38.2  53.5
#> 2      51.2   17.9   28.1    27.8     50.5     34.8   72.3      37.8  53.4
#> 3      49.0   18.1   28.5    30.6     48.8     32.8   72.0      37.4  52.2
#> 4      48.8   18.0   29.9    30.8     48.2     33.2   72.2      40.0  53.5
#> 5      50.8   17.4   29.4    30.1     48.8     36.0   72.3      37.8  54.3
#> 6      52.1   17.2   29.6    32.8     50.0     37.2   72.2      35.5  52.6
#> 7      49.9   17.3   31.6    34.0     49.0     34.5   71.4      33.7  52.3
#> 8      49.1   16.9   31.4    33.8     47.7     34.7   68.5      32.6  55.0
#> 9        NA     NA     NA      NA       NA       NA     NA        NA    NA
#> 10     48.8   17.1   33.1    38.3     47.5     34.3   68.2      32.5  51.9
#> 11       NA     NA     NA      NA       NA       NA     NA        NA    NA
#> 12     49.1   18.4   30.6    33.1     47.7     34.9   68.6      31.2  49.9
#> 13     47.9   18.8   31.4    33.1     46.7     33.7   69.9      31.9  50.1
#> 14     48.4   19.0   30.7    33.4     46.8     34.6   70.6      32.1  52.5
#> 15     47.0   20.0   33.5    36.5     46.2     32.8   70.2      29.3  51.9
#> 16     49.2   19.9   34.2    34.5     47.7     35.1   72.3      31.6  51.5
#> 17     50.2   19.9   32.8    36.9     48.9     35.5   70.8      31.4  50.3
#> 18     51.0   19.9   33.4    36.8     50.0     35.4   70.0      30.0  52.3
#> 19     50.5   21.0   33.7    38.0     48.4     36.7   71.4      31.6  55.7
#> 20       NA     NA     NA      NA       NA       NA     NA        NA    NA
#> 21     50.0   20.2   33.6    38.9     48.2     35.9   69.6      32.2  51.8
#> 22       NA     NA     NA      NA       NA       NA     NA        NA    NA
#> 23     49.6   21.9   35.7    39.9     48.1     35.0   69.8      33.8  54.9
#> 24     49.3   20.9   33.2    36.1     47.9     34.9   70.6      33.0  55.0
#> 25     51.9   20.2   33.7    37.7     50.9     36.0   70.8      33.1  56.9
#> 26     49.8   19.8   33.9    37.7     47.5     36.5   69.7      30.8  54.1
#> 27     49.0   20.7   35.3    35.3     47.2     35.5   69.6      29.7  54.5
#> 28     49.7   22.3   36.1    37.6     49.3     33.7   68.8      32.0  56.0
#> 29     49.6   21.1   34.8    40.6     49.1     33.9   68.2      30.0  54.2
#> 30     48.0   21.4   36.3    35.9     47.0     33.4   68.0      30.4  54.1
#>    blk_pct stl_pct                                            home_record
#> 1      9.5     9.4                                                  83-67
#> 2     10.7     9.7                                                  80-41
#> 3     10.0     9.2                                                  91-59
#> 4     10.8     9.0                                                  80-55
#> 5     10.8     8.8                                                  85-50
#> 6     10.0     8.5                                                  93-42
#> 7      9.6     8.7                                                  89-46
#> 8     10.0     8.7                                                  76-59
#> 9       NA      NA       Dropped: Maryland (B10) Added: Louisville (Amer)
#> 10    10.1     8.5                                                  77-58
#> 11      NA      NA Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)
#> 12    10.1     8.8                                                  74-34
#> 13    10.0     9.1                                                  58-38
#> 14    10.3     9.5                                                  63-33
#> 15    10.3     9.8                                                  66-30
#> 16    10.7    10.2                                                  61-35
#> 17    10.8     9.7                                                  58-38
#> 18     9.6     9.7                                                  63-33
#> 19    10.4    10.7                                                  60-36
#> 20      NA      NA                             Added: Boston College (BE)
#> 21    10.7    10.7                                                  55-33
#> 22      NA      NA               Added: Miami FL (BE), Virginia Tech (BE)
#> 23    10.9    11.0                                                  47-25
#> 24    10.7    10.5                                                  53-19
#> 25     8.0    10.2                                                  46-26
#> 26     9.4     9.5                                                  44-28
#> 27     9.8     9.8                                                  46-26
#> 28    10.4    10.6                                                  47-25
#> 29     9.8     9.7                                                  42-30
#> 30     8.9     9.9                                                  45-27
#>                                                      bids
#> 1                                                       5
#> 2                                                       7
#> 3                                                      5*
#> 4                                                       7
#> 5                                                       9
#> 6                                                       9
#> 7                                                       7
#> 8                                                       6
#> 9        Dropped: Maryland (B10) Added: Louisville (Amer)
#> 10                                                      6
#> 11 Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)
#> 12                                                      4
#> 13                                                      5
#> 14                                                      4
#> 15                                                      6
#> 16                                                      7
#> 17                                                      4
#> 18                                                      7
#> 19                                                      4
#> 20                             Added: Boston College (BE)
#> 21                                                      5
#> 22               Added: Miami FL (BE), Virginia Tech (BE)
#> 23                                                      6
#> 24                                                      4
#> 25                                                      4
#> 26                                                      6
#> 27                                                      3
#> 28                                                      3
#> 29                                                      5
#> 30                                                      6
#>                                                       s16
#> 1                                                       3
#> 2                                                       2
#> 3                                           No tournament
#> 4                                                       5
#> 5                                                       4
#> 6                                                       1
#> 7                                                       6
#> 8                                                       5
#> 9        Dropped: Maryland (B10) Added: Louisville (Amer)
#> 10                                                      1
#> 11 Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)
#> 12                                                      2
#> 13                                                      2
#> 14                                                      3
#> 15                                                      1
#> 16                                                      2
#> 17                                                      1
#> 18                                                      1
#> 19                                                      2
#> 20                             Added: Boston College (BE)
#> 21                                                      3
#> 22               Added: Miami FL (BE), Virginia Tech (BE)
#> 23                                                      3
#> 24                                                      2
#> 25                                                      2
#> 26                                                      2
#> 27                                                      2
#> 28                                                      2
#> 29                                                      3
#> 30                                                      2
#>                                                        f4
#> 1                                                       2
#> 2                                                       0
#> 3                                           No tournament
#> 4                                                       1
#> 5                                                       0
#> 6                                                       1
#> 7                                                       2
#> 8                                                       1
#> 9        Dropped: Maryland (B10) Added: Louisville (Amer)
#> 10                                                      0
#> 11 Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)
#> 12                                                      0
#> 13                                                      0
#> 14                                                      0
#> 15                                                      1
#> 16                                                      1
#> 17                                                      1
#> 18                                                      0
#> 19                                                      0
#> 20                             Added: Boston College (BE)
#> 21                                                      1
#> 22               Added: Miami FL (BE), Virginia Tech (BE)
#> 23                                                      2
#> 24                                                      0
#> 25                                                      1
#> 26                                                      2
#> 27                                                      1
#> 28                                                      1
#> 29                                                      1
#> 30                                                      1
#>                                                        ch
#> 1                                                       0
#> 2                                                       0
#> 3                                           No tournament
#> 4                                                       1
#> 5                                                       0
#> 6                                                       1
#> 7                                                       0
#> 8                                                       1
#> 9        Dropped: Maryland (B10) Added: Louisville (Amer)
#> 10                                                      0
#> 11 Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)
#> 12                                                      0
#> 13                                                      0
#> 14                                                      0
#> 15                                                      1
#> 16                                                      1
#> 17                                                      0
#> 18                                                      0
#> 19                                                      0
#> 20                             Added: Boston College (BE)
#> 21                                                      1
#> 22               Added: Miami FL (BE), Virginia Tech (BE)
#> 23                                                      0
#> 24                                                      0
#> 25                                                      1
#> 26                                                      1
#> 27                                                      0
#> 28                                                      0
#> 29                                                      0
#> 30                                                      0
#>                                          reg_season_champ
#> 1                                                  Duke 2
#> 2                                              Virginia 4
#> 3                                          Florida St. 2*
#> 4                                               2-way tie
#> 5                                              Virginia 1
#> 6                                        North Carolina 1
#> 7                                        North Carolina 1
#> 8                                              Virginia 2
#> 9        Dropped: Maryland (B10) Added: Louisville (Amer)
#> 10                                             Virginia 1
#> 11 Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)
#> 12                                             Miami FL 2
#> 13                                       North Carolina 1
#> 14                                       North Carolina 2
#> 15                                              2-way tie
#> 16                                       North Carolina 1
#> 17                                       North Carolina 1
#> 18                                              2-way tie
#> 19                                                 Duke 1
#> 20                             Added: Boston College (BE)
#> 21                                       North Carolina 1
#> 22               Added: Miami FL (BE), Virginia Tech (BE)
#> 23                                                 Duke 1
#> 24                                          Wake Forest 2
#> 25                                             Maryland 1
#> 26                                              2-way tie
#> 27                                                 Duke 1
#> 28                                                 Duke 1
#> 29                                                 Duke 1
#> 30                                                 Duke 2
#>                                             tourney_champ
#> 1                                        Virginia Tech 11
#> 2                                          Georgia Tech 9
#> 3                                                        
#> 4                                                  Duke 1
#> 5                                              Virginia 1
#> 6                                                  Duke 2
#> 7                                        North Carolina 1
#> 8                                            Notre Dame 3
#> 9        Dropped: Maryland (B10) Added: Louisville (Amer)
#> 10                                             Virginia 1
#> 11 Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)
#> 12                                             Miami FL 2
#> 13                                          Florida St. 3
#> 14                                                 Duke 1
#> 15                                                 Duke 1
#> 16                                                 Duke 2
#> 17                                       North Carolina 1
#> 18                                       North Carolina 1
#> 19                                                 Duke 1
#> 20                             Added: Boston College (BE)
#> 21                                                 Duke 1
#> 22               Added: Miami FL (BE), Virginia Tech (BE)
#> 23                                             Maryland 4
#> 24                                                 Duke 3
#> 25                                                 Duke 1
#> 26                                                 Duke 1
#> 27                                                 Duke 1
#> 28                                                 Duke 1
#> 29                                       North Carolina 1
#> 30                                       North Carolina 1
#>                                                 best_team
#> 1                                                  Duke 2
#> 2                                           Florida St. 4
#> 3                                                 Duke 3*
#> 4                                              Virginia 1
#> 5                                              Virginia 1
#> 6                                        North Carolina 1
#> 7                                        North Carolina 1
#> 8                                                  Duke 1
#> 9        Dropped: Maryland (B10) Added: Louisville (Amer)
#> 10                                             Virginia 1
#> 11 Added: Notre Dame (BE), Pittsburgh (BE), Syracuse (BE)
#> 12                                                 Duke 2
#> 13                                       North Carolina 1
#> 14                                                 Duke 1
#> 15                                                 Duke 1
#> 16                                       North Carolina 1
#> 17                                       North Carolina 1
#> 18                                       North Carolina 1
#> 19                                                 Duke 1
#> 20                             Added: Boston College (BE)
#> 21                                       North Carolina 1
#> 22               Added: Miami FL (BE), Virginia Tech (BE)
#> 23                                                 Duke 1
#> 24                                                 Duke 3
#> 25                                                 Duke 1
#> 26                                                 Duke 1
#> 27                                                 Duke 1
#> 28                                                 Duke 1
#> 29                                                 Duke 1
#> 30                                       North Carolina 1
# }