Load hoopR NBA play-by-play

helper that loads multiple seasons from the data repo either into memory or writes it into a db using some forwarded arguments in the dots

Usage

load_nba_pbp(
  seasons = most_recent_nba_season(),
  ...,
  dbConnection = NULL,
  tablename = NULL
)

Arguments

seasons

A vector of 4-digit years associated with given NBA seasons. (Min: 2002)

...

Additional arguments passed to an underlying function that writes the season data into a database (used by update_nba_db()).

dbConnection

A DBIConnection object, as returned by

tablename

The name of the play by play data table within the database

Value

Returns a tibble

See also

Examples

# \donttest{
  load_nba_pbp(seasons = most_recent_nba_season())
#> ── ESPN NBA Play-by-Play Information from hoopR data repository ────────────────
#> ℹ Data updated: 2023-03-27 10:14:20 UTC
#> # A tibble: 528,919 × 56
#>             id sequence_number type_id type_text     text  away_score home_score
#>          <dbl> <chr>           <chr>   <chr>         <chr>      <int>      <int>
#>  1  4014680164 4               615     Jumpball      Joel…          0          0
#>  2  4014680167 7               137     Turnaround F… Jayl…          0          0
#>  3  4014680169 9               156     Offensive Re… P.J.…          0          0
#>  4 40146801610 10              87      Out of Bound… P.J.…          0          0
#>  5 40146801611 11              145     Driving Floa… Marc…          0          2
#>  6 40146801612 12              144     Driving Floa… Tobi…          0          2
#>  7 40146801613 13              156     Offensive Re… Tobi…          0          2
#>  8 40146801614 14              94      Tip Shot      Tobi…          2          2
#>  9 40146801615 15              92      Jump Shot     Jays…          2          5
#> 10 40146801617 17              92      Jump Shot     Joel…          2          5
#> # ℹ 528,909 more rows
#> # ℹ 49 more variables: period_number <int>, period_display_value <chr>,
#> #   clock_display_value <chr>, scoring_play <lgl>, score_value <int>,
#> #   team_id <chr>, participants_0_athlete_id <chr>,
#> #   participants_1_athlete_id <chr>, participants_2_athlete_id <chr>,
#> #   wallclock <chr>, shooting_play <lgl>, coordinate_x <int>,
#> #   coordinate_y <int>, season <int>, season_type <int>, away_team_id <int>, …
# }