Get men's college basketball NET rankings for the current date from the NCAA website

Get men's college basketball NET rankings for the current date from the NCAA website

ncaa_mbb_NET_rankings()

Value

Returns a tibble

Author

Saiem Gilani

Examples

# Get current NCAA NET rankings
ncaa_mbb_NET_rankings()

#> # A tibble: 347 x 12
#>     rank previous school     conference record road  neutral home  quad_1 quad_2
#>    <int>    <int> <chr>      <chr>      <chr>  <chr> <chr>   <chr> <chr>  <chr> 
#>  1     1        1 Gonzaga    WCC        31-1   7-0   12-1    12-0  12-1   6-0   
#>  2     2        2 Baylor     Big 12     28-2   7-1   10-1    11-0  13-2   3-0   
#>  3     3        4 Michigan   Big Ten    23-5   6-2   4-2     13-1  10-3   4-2   
#>  4     4        3 Illinois   Big Ten    24-7   9-3   4-2     11-2  12-5   5-2   
#>  5     5        5 Houston    AAC        27-4   5-3   8-1     14-0  6-2    5-1   
#>  6     6       19 Southern … Pac-12     25-8   7-3   5-3     13-2  9-7    7-1   
#>  7     7        6 Iowa       Big Ten    22-9   6-4   2-3     14-2  7-7    5-2   
#>  8     8        7 Alabama    SEC        26-7   7-3   7-3     12-1  9-5    8-1   
#>  9     9       15 Colorado   Pac-12     23-9   6-6   6-2     11-1  8-6    5-2   
#> 10    10       10 Loyola Ch… MVC        24-5   8-3   5-2     11-0  4-3    4-2   
#> # … with 337 more rows, and 2 more variables: quad_3 <chr>, quad_4 <chr>