Get men's college basketball NET rankings for the current date from the NCAA website

Get men's college basketball NET rankings for the current date from the NCAA website

Usage

ncaa_mbb_NET_rankings()

Value

Returns a tibble

Author

Saiem Gilani

Examples

# Get current NCAA NET rankings
# \donttest{
  try(ncaa_mbb_NET_rankings())
#> ── NCAA MBB NET Rankings Information from ESPN.com ────────────── hoopR 1.8.0 ──
#> ℹ Data updated: 2022-06-16 03:23:00 UTC
#> # A tibble: 358 × 12
#>     rank previous school     conference record road  neutral home  quad_1 quad_2
#>    <int>    <int> <chr>      <chr>      <chr>  <chr> <chr>   <chr> <chr>  <chr> 
#>  1     1        1 Gonzaga    WCC        28-4   5-1   7-3     16-0  11-4   2-0   
#>  2     2        3 Houston    AAC        32-6   7-3   9-2     16-1  5-5    10-1  
#>  3     3        6 Kansas     Big 12     34-6   6-4   12-1    16-1  16-5   10-1  
#>  4     4        2 Arizona    Pac-12     33-4   9-3   7-1     17-0  7-4    9-0   
#>  5     5        4 Baylor     Big 12     27-7   8-3   4-2     15-2  11-6   8-1   
#>  6     6        8 Villanova  Big East   30-8   10-5  8-2     12-1  12-7   8-1   
#>  7     7        7 Tennessee  SEC        27-8   6-5   5-3     16-0  11-8   5-0   
#>  8     8        9 Texas Tech Big 12     27-10  3-7   6-3     18-0  10-10  7-0   
#>  9     9        5 Kentucky   SEC        26-8   6-5   2-3     18-0  8-7    6-0   
#> 10    10       12 Duke       ACC        32-7   9-2   8-2     15-3  11-4   5-3   
#> # … with 348 more rows, and 2 more variables: quad_3 <chr>, quad_4 <chr>
# }